refactor: AppDelegate is the composition root and nothing else
It was 867 lines holding every module's wiring, and the line-limit
baseline recorded it as debt to be paid down someday. It is now 95: the
list of what the app is made of, the order it comes up in, and the check
for a second copy.

Everything else moved into an AppDelegate+<Module>.swift beside it,
which is the convention AGENTS.md already prescribed and that
+Awake, +Active, +Zones and +Shell were already following. New files:
Wiring, Model, DragSnap, Mouse, Update, Server, Screens, Login and
Actions. perform() and the capture helpers joined +Shortcuts and
+Capture, which already owned those modules.

Nothing changed but where the code lives. The stored properties lost
`private` because Swift only lets a same-file extension see that, and
the class is now split on purpose; the comment at the top of the class
says so.

AppDelegate and Router are both out of scripts/line-limit-baseline
rather than lowered in it, which leaves seven entries where there were
ten, and no source file in App/Sources over the limit.
