+++ App/Sources/plonk/AppDelegate.swift
@@ -6,37 +6,40 @@ import SwiftUI
 final class AppDelegate: NSObject, NSApplicationDelegate {
     static let issuesURL = "https://github.com/ostapondo/plonk/issues/new"
 
+    // Split across the AppDelegate+* files, one per module, so nothing here is
+    // private: Swift only lets an extension in the same file see that. What is
+    // listed here is the whole of what the app is made of.
     let store = ConfigStore()
-    private let windows = WindowManager()
+    let windows = WindowManager()
     let awake = AwakeManager()
     let active = ActiveManager()
     let agents = AgentRegistry()
-    private let eventBroadcaster = EventBroadcaster()
+    let eventBroadcaster = EventBroadcaster()
     let voice = VoiceManager()
     let hotkeys = HotkeyManager()
-    private let updates = UpdateManager()
+    let updates = UpdateManager()
     let model = AppModel()
-    private let snapMemory = SnapMemory()
+    let snapMemory = SnapMemory()
     lazy var commands = WindowCommands(windows: windows, memory: snapMemory)
-    private lazy var launcher = WorkspaceLauncher(windows: windows)
+    lazy var launcher = WorkspaceLauncher(windows: windows)
     lazy var presenter = WindowPresenter(model: model)
     var statusMenu: StatusMenuController!
     var dragSnap: DragSnapManager!
-    private var grabMove: GrabMove!
-    private var newWindows: NewWindowWatcher!
-    private let mouse = MouseTools()
-    private let crops = CropAndLock()
-    private var guidePanel: ShortcutGuidePanel?
-    private var guideLoading = false
+    var grabMove: GrabMove!
+    var newWindows: NewWindowWatcher!
+    let mouse = MouseTools()
+    let crops = CropAndLock()
+    var guidePanel: ShortcutGuidePanel?
+    var guideLoading = false
     /// How many captures are in flight; see hideOwnWindows.
-    private var captureDepth = 0
+    var captureDepth = 0
     var router: Router!
-    private var server: ControlServer?
+    var server: ControlServer?
     var previewToken = 0
-    private var screenSettleWork: DispatchWorkItem?
+    var screenSettleWork: DispatchWorkItem?
     /// Display UUIDs seen at the last screen-parameter change, so a Dock
     /// resize can be told apart from a monitor being plugged in.
-    private var knownDisplays: Set<String> = []
+    var knownDisplays: Set<String> = []
 
     func applicationDidFinishLaunching(_ notification: Notification) {
         // Before anything is on screen: a second copy would add its own menu bar
@@ -87,781 +90,5 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
         let ownPID = ProcessInfo.processInfo.processIdentifier
         return NSRunningApplication.runningApplications(withBundleIdentifier: id)
             .first { $0.processIdentifier != ownPID && !$0.isTerminated }
-    }
-
-    // MARK: - Wiring
-
-    /// One place decides what Plonk will not touch, and everything that moves
-    /// a window on its own asks it.
-    private func isExcluded(_ app: NSRunningApplication) -> Bool {
-        AppExclusions.matches(name: app.localizedName ?? "", bundleID: app.bundleIdentifier,
-                              patterns: store.config.excludedApps)
-    }
-
-    private func setupCommands() {
-        commands.zonesForScreen = { [weak self] index in
-            guard let self else { return [] }
-            return store.config.zones(forKeys: ScreenIdentity.keys(forIndex: index))
-        }
-        commands.isExcluded = { [weak self] app in self?.isExcluded(app) ?? false }
-        commands.announce = { HUD.shared.show($0) }
-        commands.zoneGap = { [weak self] in CGFloat(self?.store.config.zoneGap ?? 0) }
-    }
-
-    private func setupPresenter() {
-        presenter.shotFolder = { [weak self] in self?.store.config.shotFolder ?? "~/Desktop" }
-        presenter.onShotFinished = { [weak self] copied, path in self?.finishShot(copied: copied, path: path) }
-        presenter.onCancelWorkspaceLaunch = { [weak self] in self?.launcher.cancel() }
-        presenter.onZonePickerClosed = { [weak self] in self?.clearPreview() }
-        presenter.onFullscreenEditorClosed = { [weak self] in self?.clearPreview() }
-    }
-
-    private func setupLauncher() {
-        launcher.onProgress = { [weak self] name, statuses in
-            guard let self else { return }
-            model.launchingWorkspace = name
-            model.launchStatuses = statuses
-            presenter.showWorkspaceLaunch()
-        }
-        launcher.onFinished = { [weak self] _, statuses in
-            guard let self else { return }
-            model.launchStatuses = statuses
-            model.launchingWorkspace = nil
-            // Anything that failed stays up until the user has read it.
-            guard statuses.allSatisfy({ $0.state != .pending && !$0.isFailure }) else { return }
-            DispatchQueue.main.asyncAfter(deadline: .now() + 1.2) { [weak self] in
-                guard self?.model.launchingWorkspace == nil else { return }
-                self?.presenter.closeWorkspaceLaunch()
-            }
-        }
-    }
-
-    func perform(_ action: HotkeyAction) {
-        switch action {
-        case .showZones:
-            dragSnap.previewZones()
-        case .captureRegion:
-            runCapture(.region, openEditor: true)
-        case .captureText:
-            captureText()
-        case .voice:
-            voice.beginCapture()
-        case .unsnap:
-            commands.unsnap()
-        case .cycleZone:
-            commands.cycleZone(backwards: false)
-        case .cycleZoneBack:
-            commands.cycleZone(backwards: true)
-        case .findCursor:
-            mouse.flashSpotlight()
-        case .jumpCursor:
-            if !mouse.jumpToNextScreen() { HUD.shared.show(.hudOneScreenOnly) }
-        case .cropLive:
-            pinCrop(live: true)
-        case .cropStill:
-            pinCrop(live: false)
-        case .ruler:
-            startRuler()
-        case .shortcutGuide:
-            toggleShortcutGuide()
-        case .commandPalette:
-            openCommandPalette()
-        case .zoneSetPalette:
-            openZoneSetPalette()
-        default:
-            if let number = action.zoneNumber {
-                commands.snap(toZone: number)
-            } else if let number = action.layoutNumber {
-                commands.applyZoneSet(number: number, named: model.zoneSetNames) { [weak self] name, screen in
-                    self?.assignZoneSet(name, toScreen: screen)
-                }
-            } else if let direction = action.focusDirection {
-                commands.moveFocus(direction)
-            } else if let preset = action.preset {
-                commands.apply(preset)
-            }
-        }
-    }
-
-    private func setupVoice() {
-        voice.onPartial = { text in
-            HUD.shared.show(text.isEmpty ? String(localized: .hudListening) : text)
-        }
-        voice.onError = { message in HUD.shared.show(message) }
-        voice.onTranscript = { [weak self] text in
-            guard let self else { return }
-            // The dozen things people say most run here, with no agent, no
-            // round trip and no network. Everything else takes the long road.
-            if store.config.voiceLocalCommands,
-               let command = VoiceCommand.parse(text, workspaces: model.workspaceNames) {
-                run(command)
-                HUD.shared.show(.hudRan(String(localized: command.announcement)))
-                return
-            }
-            let response = router.dispatch(prompt: text)
-            if let error = response.json["error"] as? String {
-                HUD.shared.show(error)
-            } else {
-                let agent = (response.json["agent"] as? String) ?? "agent"
-                HUD.shared.show(.hudSentTo(agent, text))
-            }
-        }
-    }
-
-    /// Wiring only: what drag snapping asks the app for, and what it reports
-    /// back. The settings it reads are applied separately, because they change
-    /// while the app is running and this does not.
-    private func setupDragSnap() {
-        dragSnap = DragSnapManager(windows: windows)
-        dragSnap.zonesForScreen = { [weak self] index in
-            guard let self else { return [] }
-            return store.config.zones(forKeys: ScreenIdentity.keys(forIndex: index))
-        }
-        dragSnap.appearance = { [weak self] in self?.zoneAppearance ?? ZoneAppearance() }
-        dragSnap.isExcluded = { [weak self] app in self?.isExcluded(app) ?? false }
-        dragSnap.onSnap = { [weak self] window, before, frac, screenIndex in
-            guard let self else { return }
-            snapMemory.record(window, wasAt: before, placedAt: frac,
-                              screenUUID: ScreenIdentity.uuid(forIndex: screenIndex),
-                              zoneIndex: zoneIndex(of: frac, onScreen: screenIndex),
-                              appKey: windows.app(ofWindow: window)?.bundleIdentifier)
-        }
-        applyDragSnapSettings()
-        dragSnap.start()
-    }
-
-    private func applyDragSnapSettings() {
-        dragSnap.enabled = store.config.dragSnapEnabled
-        dragSnap.requireModifier = store.config.zonesRequireShift
-        dragSnap.modifierFlag = Self.modifierFlag(store.config.zonesModifier)
-        dragSnap.showOnAllMonitors = store.config.zonesOnAllMonitors
-        dragSnap.edgeSpanPoints = store.config.zoneEdgeSpanPoints
-    }
-
-    var zoneAppearance: ZoneAppearance {
-        ZoneAppearance(gap: CGFloat(store.config.zoneGap),
-                       opacity: CGFloat(store.config.zoneOpacity),
-                       // Falls back to the app's accent, which itself falls back
-                       // to the system one, so the overlay is part of the theme.
-                       color: ZoneAppearance.color(fromHex: store.config.zoneColorHex)
-                           ?? ZoneAppearance.color(fromHex: store.config.appearance.accentHex),
-                       showNumbers: store.config.zoneNumbersVisible)
-    }
-
-    private func setupGrabMove() {
-        grabMove = GrabMove(windows: windows)
-        grabMove.isExcluded = { [weak self] app in self?.isExcluded(app) ?? false }
-        // A grab is a drag as far as zones are concerned, so it goes through
-        // the same overlay and the same drop rules.
-        grabMove.onGrabBegan = { [weak self] window, frame in
-            self?.dragSnap.beginExternalDrag(window: window, startFrame: frame)
-        }
-        grabMove.onGrabMoved = { [weak self] in self?.dragSnap.updateExternalDrag() }
-        grabMove.onGrabResized = { [weak self] window, startFrame in
-            guard let self, let placed = windows.fraction(ofWindow: window) else { return }
-            snapMemory.record(window, wasAt: startFrame, placedAt: placed.frac,
-                              screenUUID: ScreenIdentity.uuid(forIndex: placed.screenIndex),
-                              appKey: windows.app(ofWindow: window)?.bundleIdentifier)
-            router?.changes.bump("windows")
-        }
-        grabMove.onGrabEnded = { [weak self] window, startFrame in
-            guard let self else { return }
-            // A grab that landed in a zone is recorded by the zone drop. One
-            // that did not is still a move Plonk made, so it is remembered
-            // too — otherwise the shortcut that puts a window back would have
-            // nothing to put back after a free drag.
-            if !dragSnap.endExternalDrag(), let placed = windows.fraction(ofWindow: window) {
-                snapMemory.record(window, wasAt: startFrame, placedAt: placed.frac,
-                                  screenUUID: ScreenIdentity.uuid(forIndex: placed.screenIndex))
-            }
-            router?.changes.bump("windows")
-        }
-        applyGrabMoveSettings()
-    }
-
-    /// An event tap that can swallow clicks has no business existing while the
-    /// feature is off, so the tap follows the setting rather than the launch.
-    private func applyGrabMoveSettings() {
-        grabMove.enabled = store.config.grabMoveEnabled
-        grabMove.modifierFlag = Self.modifierFlag(store.config.grabMoveModifier)
-        grabMove.allowResize = store.config.grabMoveResize
-        grabMove.showGeometry = store.config.grabMoveShowGeometry
-        if store.config.grabMoveEnabled {
-            grabMove.start()
-        } else {
-            grabMove.stop()
-        }
-    }
-
-    /// Which numbered zone a dropped fraction corresponds to, so editing the
-    /// set later can move the window with its number. Nil for a span or an
-    /// edge snap, which match no single zone.
-    private func zoneIndex(of frac: FracRect, onScreen index: Int) -> Int? {
-        let zones = store.config.zones(forKeys: ScreenIdentity.keys(forIndex: index))
-        return zones.firstIndex {
-            abs($0.x - frac.x) < 0.001 && abs($0.y - frac.y) < 0.001
-                && abs($0.w - frac.w) < 0.001 && abs($0.h - frac.h) < 0.001
-        }
-    }
-
-    private func setupNewWindows() {
-        newWindows = NewWindowWatcher(windows: windows)
-        newWindows.enabled = store.config.placeNewWindows
-        newWindows.isExcluded = { [weak self] app in self?.isExcluded(app) ?? false }
-        newWindows.zoneGap = { [weak self] in CGFloat(self?.store.config.zoneGap ?? 0) }
-        newWindows.placement = { [weak self] app in
-            guard let self, let key = app.bundleIdentifier,
-                  let habit = snapMemory.habit(ofApp: key) else { return nil }
-            // The display is named by UUID, so a habit formed on a monitor that
-            // has since been unplugged simply does not apply.
-            guard let uuid = habit.screenUUID, let index = ScreenIdentity.index(forUUID: uuid) else { return nil }
-            // A numbered zone is followed by number, so the habit survives the
-            // set being edited; anything else falls back to the raw fraction.
-            let zones = store.config.zones(forKeys: ScreenIdentity.keys(forIndex: index))
-            if let zone = habit.zoneIndex, zones.indices.contains(zone) {
-                return (zones[zone].frac, index)
-            }
-            return (habit.frac, index)
-        }
-        newWindows.onPlaced = { [weak self] app, window, before, frac, screenIndex in
-            guard let self else { return }
-            snapMemory.record(window, wasAt: before, placedAt: frac,
-                              screenUUID: ScreenIdentity.uuid(forIndex: screenIndex),
-                              zoneIndex: zoneIndex(of: frac, onScreen: screenIndex),
-                              appKey: app.bundleIdentifier)
-            router?.changes.bump("windows")
-        }
-        newWindows.start()
-    }
-
-    /// Reads the front app's menus and floats them. Pressing the key again
-    /// closes it, which is how a reference panel should behave.
-    private func toggleShortcutGuide() {
-        if let panel = guidePanel {
-            panel.close()
-            return
-        }
-        // Reading a big app's menus takes a moment; a second press inside that
-        // window would build a second panel and orphan the first.
-        guard !guideLoading else { return }
-        guard windows.isTrusted else {
-            HUD.shared.show(.hudGuideNeedsAccessibility)
-            return
-        }
-        guard let app = NSWorkspace.shared.frontmostApplication else { return }
-        let name = app.localizedName ?? String(localized: .guideThisApp)
-        guideLoading = true
-        ShortcutGuide.read(for: app) { [weak self] items in
-            guard let self else { return }
-            guideLoading = false
-            // The guide describes whatever was in front when it was asked for,
-            // so a slow app cannot end up labelled with the wrong name.
-            let panel = ShortcutGuidePanel(appName: name, items: items)
-            panel.onClose = { [weak self] in self?.guidePanel = nil }
-            guidePanel = panel
-            panel.orderFrontRegardless()
-        }
-    }
-
-    /// Drag out a region and float it above everything, live or frozen.
-    private func pinCrop(live: Bool) {
-        let report: (Result<String, Error>) -> Void = { result in
-            if case .failure(let error) = result {
-                HUD.shared.show(error.localizedDescription)
-            }
-        }
-        // Plonk's own windows would otherwise be in the shot, and the picker
-        // has to be the only thing on top.
-        let hidden = hideOwnWindows()
-        let finish: (Result<String, Error>) -> Void = { [weak self] result in
-            self?.showOwnWindows(hidden)
-            report(result)
-        }
-        if live {
-            crops.createLive(completion: finish)
-        } else {
-            crops.createStill(completion: finish)
-        }
-    }
-
-    private func setupMouseTools() {
-        applyMouseSettings()
-    }
-
-    /// Same rule as grab-and-move: no tap unless something needs one. Finding
-    /// the pointer is a shortcut and draws without watching anything.
-    func applyMouseSettings() {
-        mouse.highlightEnabled = store.config.highlightClicksEnabled
-        mouse.crosshairsEnabled = store.config.crosshairsEnabled
-        mouse.tint = zoneAppearance.tint
-        if store.config.highlightClicksEnabled || store.config.crosshairsEnabled {
-            mouse.start()
-        } else {
-            mouse.stop()
-        }
-    }
-
-    private func setupUpdates() {
-        updates.onChange = { [weak self] in
-            guard let self else { return }
-            refreshUpdateModel()
-            refreshStatusMenu()
-            router?.changes.bump("update")
-        }
-        // Progress moves the bar and nothing else: agents are told about an
-        // update starting and finishing, not about every chunk of it.
-        updates.onProgress = { [weak self] in
-            self?.model.updateProgress = self?.updates.progress ?? 0
-        }
-        updates.automatic = store.config.updateCheckAutomatically
-        statusMenu.updateVersion = { [weak self] in self?.updates.available?.version.text }
-        statusMenu.onOpenUpdate = { [weak self] in self?.openPage("update") }
-
-        router.updateState = { [weak self] in
-            guard let self else { return [:] }
-            var state: [String: Any] = [
-                "installed": UpdateManager.currentVersionText,
-                "available": updates.available != nil,
-                "phase": updates.phase.rawValue,
-                "status": String(localized: updates.status),
-                "automatic": store.config.updateCheckAutomatically,
-            ]
-            if let release = updates.available {
-                state["latest"] = release.version.text
-                state["notes"] = release.notes
-                state["page"] = release.pageURL.absoluteString
-            }
-            return state
-        }
-        router.checkForUpdates = { [weak self] in self?.updates.check() }
-        router.installUpdate = { [weak self] in
-            // The reply has to reach the caller before the process goes away.
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) { self?.updates.install() }
-        }
-        updates.start()
-    }
-
-    private func setupServer() {
-        router = Router(store: store, windows: windows, awake: awake, active: active, agents: agents)
-        router.changes.onEvent = { [weak self] rev, what in
-            self?.eventBroadcaster.broadcast(rev: rev, what: what)
-        }
-        router.attachEvents = { [weak self] conn, rev in
-            self?.eventBroadcaster.attach(conn, rev: rev)
-        }
-        // Every source of change funnels into the bus at its own choke point,
-        // so no caller has to remember to announce itself.
-        //
-        // Config is the widest of them: a settings row, an agent over the API
-        // and an imported Rectangle setup all end at ConfigStore.update, so
-        // applying a change is hung there rather than at each of the three.
-        // Set here, once the managers applyConfig reaches all exist — writes
-        // before this point are startup reading its own defaults back.
-        store.didMutate = { [weak self] in
-            guard let self else { return }
-            applyConfig()
-            router.changes.bump("config")
-        }
-        windows.onDidPlace = { [weak self] in
-            self?.router.changes.bump("windows")
-            self?.markGettingStarted { $0.sawFirstSnap = true }
-        }
-        agents.onChange = { [weak self] in
-            self?.refreshAgentModel()
-            self?.router.changes.bump("agents")
-        }
-        // Only the part that is not a config change: an agent editing zones
-        // shows them, so the user sees what it did.
-        router.didChangeZones = { [weak self] in self?.dragSnap.previewZones() }
-        setupShotRoutes()
-        router.launchWorkspace = { [weak self] name, workspace, screen, done in
-            guard let self else { return done([["ok": false, "error": "shutting down"]]) }
-            launcher.launch(workspace, named: name, onScreen: screen, completion: done)
-        }
-        router.runAdapter = { [weak self] adapter, prompt in
-            self?.launchAdapter(adapter, prompt: prompt)
-        }
-
-        let token = APIToken.loadOrCreate()
-        if token == nil {
-            model.apiWarning = String(localized: .warningNoToken(APIToken.url().path))
-        }
-        let server = ControlServer(token: token) { [weak self] request, respond in
-            guard let self else { return respond(.failed("shutting down")) }
-            router.handle(request, respond: respond)
-        }
-        server.onUnavailable = { [weak self] message in self?.model.apiWarning = message }
-        do {
-            try server.start()
-        } catch {
-            NSLog("Plonk: failed to start control server: \(error)")
-            model.apiWarning = String(localized: .warningApiDidNotStart)
-        }
-        self.server = server
-    }
-
-    // MARK: - Model
-
-    /// Bundle version; empty when running unbundled through `swift run`.
-    private var appVersion: String {
-        Bundle.main.infoDictionary?["CFBundleShortVersionString"] as? String ?? ""
-    }
-
-    /// The whole model, for a window about to open. Everything config decides
-    /// goes through applyConfig; what is left is fixed for the run or comes
-    /// from macOS.
-    func refreshModel() {
-        model.appVersion = appVersion
-        model.configWarning = store.loadFailure
-        model.supportedTextLanguages = TextExtractor.supportedLanguages
-        model.settingsGroups = SettingsPages.groups
-        model.settingsPages = SettingsPages.all
-        refreshPermissions()
-        applyConfig()
-    }
-
-    /// Everything a config change has to reach, in one place, run after every
-    /// write whoever made it: a settings row, an agent over the API, a
-    /// Rectangle setup being imported, a hotkey being rebound.
-    ///
-    /// Each manager takes the whole config rather than the field that moved,
-    /// so none of them has to be told which setting it cares about. That costs
-    /// a few assignments per change and buys the guarantee that a saved
-    /// setting is an applied one.
-    func applyConfig() {
-        let config = store.config
-        model.config = config
-
-        applyAppearance()
-        hotkeys.setEnabled(config.hotkeysEnabled)
-        hotkeys.bindings = config.resolvedHotkeys
-        applyDragSnapSettings()
-        applyGrabMoveSettings()
-        applyMouseSettings()
-        awake.apply(config)
-        active.apply(config)
-        newWindows.enabled = config.placeNewWindows
-        updates.automatic = config.updateCheckAutomatically
-
-        // What the config implies rather than states: the names in it, sorted
-        // for the lists, and what each manager makes of its new settings.
-        refreshAwakeModel()
-        refreshActiveModel()
-        refreshWorkspaceModel()
-        refreshZoneModel()
-        refreshHotkeyModel()
-        refreshAgentModel()
-        refreshUpdateModel()
-    }
-
-    /// Preflight does not prompt, so it is safe to poll whenever a window opens.
-    func refreshPermissions() {
-        model.accessibilityGranted = windows.isTrusted
-        model.screenRecordingGranted = CGPreflightScreenCaptureAccess()
-    }
-
-    private func refreshUpdateModel() {
-        model.updateAvailableVersion = updates.available?.version.text ?? ""
-        model.updateNotes = updates.available?.notes ?? ""
-        model.updateStatus = String(localized: updates.status)
-        model.updatePhase = updates.phase.rawValue
-        model.updateProgress = updates.progress
-    }
-
-    private func refreshAgentModel() {
-        model.connectedAgents = agents.onlineNames()
-        if !model.connectedAgents.isEmpty {
-            markGettingStarted { $0.sawFirstAgent = true }
-        }
-    }
-
-    /// Tick a Getting Started step, at most once. Config is written to disk on
-    /// every update, and both of these fire from paths that run constantly —
-    /// every placement, every agent poll — so the guard is the point.
-    private func markGettingStarted(_ change: @escaping (inout Config) -> Void) {
-        DispatchQueue.main.async { [weak self] in
-            guard let self else { return }
-            var updated = self.store.config
-            change(&updated)
-            guard updated.sawFirstSnap != self.store.config.sawFirstSnap
-                || updated.sawFirstAgent != self.store.config.sawFirstAgent else { return }
-            self.store.update(change)
-        }
-    }
-
-    private func refreshWorkspaceModel() {
-        model.workspaceNames = store.config.workspaces.keys.sorted()
-        statusMenu.workspaceNames = model.workspaceNames
-    }
-
-    func refreshZoneModel() {
-        let custom = store.config.zoneSets.keys.sorted()
-        model.customZoneSetNames = custom
-        model.zoneSetNames = custom + BuiltinZoneSets.all.keys.sorted().filter { !custom.contains($0) }
-        model.zoneSets = BuiltinZoneSets.all.merging(store.config.zoneSets) { _, user in user }
-        model.screenCount = NSScreen.screens.count
-        model.screenDescriptions = NSScreen.screens.map { "\(Int($0.frame.width)) × \(Int($0.frame.height))" }
-        model.screenAssignments = NSScreen.screens.indices.reduce(into: [:]) { result, index in
-            result[index] = store.config.zoneAssignment(forKeys: ScreenIdentity.keys(forIndex: index))
-        }
-    }
-
-    func refreshStatusMenu() {
-        statusMenu.refresh(
-            icon: awake.isOn ? StatusIcon.awake : StatusIcon.idle,
-            tooltip: String(localized: .menuTooltip(String(localized: awake.statusText))),
-            dimmed: awake.requested && !awake.isOn
-        )
-    }
-
-    func clearPreview() {
-        dragSnap.hidePreviews()
-        model.previewedZoneSet = nil
-    }
-
-    @objc private func screensChanged() {
-        dragSnap.screensChanged()
-        model.previewedZoneSet = nil
-        refreshZoneModel()
-
-        // This notification also fires for a resolution change, a display
-        // waking, and the Dock being resized or auto-hidden. Only a display
-        // actually arriving or leaving should move anybody's windows.
-        let displays = Self.attachedDisplays()
-        defer { knownDisplays = displays }
-        guard store.config.restoreZonesOnScreenChange, displays != knownDisplays else { return }
-
-        // macOS sends this before the apps behind those windows have caught up,
-        // and several times while a display wakes. Settling first, then placing
-        // once, is the difference between windows landing and windows fighting.
-        screenSettleWork?.cancel()
-        let work = DispatchWorkItem { [weak self] in self?.commands.restorePlacements() }
-        screenSettleWork = work
-        DispatchQueue.main.asyncAfter(deadline: .now() + 1.5, execute: work)
-    }
-
-    /// Grab-and-move, the pointer tools and the new-window watcher all need
-    /// Accessibility to arm, and on a first run it has just been asked for and
-    /// not yet given. Rather than making each of them dead until the next
-    /// launch, this waits for the grant and starts them then. Drag snapping
-    /// does not need it because it re-checks on every event.
-    private func watchForAccessibility() {
-        guard !windows.isTrusted else { return }
-        let timer = Timer(timeInterval: 2, repeats: true) { [weak self] timer in
-            guard let self else {
-                timer.invalidate()
-                return
-            }
-            guard windows.isTrusted else { return }
-            timer.invalidate()
-            applyGrabMoveSettings()
-            applyMouseSettings()
-            newWindows.start()
-            refreshPermissions()
-        }
-        RunLoop.main.add(timer, forMode: .common)
-    }
-
-    private static func attachedDisplays() -> Set<String> {
-        Set(NSScreen.screens.indices.compactMap { ScreenIdentity.uuid(forIndex: $0) })
-    }
-
-    private static func modifierFlag(_ name: String) -> NSEvent.ModifierFlags {
-        switch name {
-        case "option": return .option
-        case "control": return .control
-        case "command": return .command
-        default: return .shift
-        }
-    }
-
-    // MARK: - Launch at login
-
-    private var isLaunchAtLoginEnabled: Bool {
-        SMAppService.mainApp.status == .enabled
-    }
-
-    /// Registers the app bundle as a login item. Fails silently when the app
-    /// runs unbundled (swift run), where there is nothing to register.
-    private func applyLaunchAtLogin(_ enabled: Bool) {
-        do {
-            if enabled {
-                if SMAppService.mainApp.status != .enabled { try SMAppService.mainApp.register() }
-            } else if SMAppService.mainApp.status == .enabled {
-                try SMAppService.mainApp.unregister()
-            }
-        } catch {
-            NSLog("Plonk: login item update failed: \(error)")
-        }
-    }
-
-    // MARK: - Screenshot
-
-    func runCapture(_ mode: CaptureMode, openEditor: Bool,
-                            completion: @escaping (NSImage?) -> Void = { _ in }) {
-        let hidden = hideOwnWindows()
-        ScreenshotManager.capture(mode) { [weak self] image in
-            self?.showOwnWindows(hidden)
-            if let image, openEditor { self?.presenter.showShotEditor(image: image) }
-            completion(image)
-        }
-    }
-
-    /// Everything Plonk has on screen, ordered out so it cannot end up in a
-    /// capture: settings and editors, but also the crosshairs and any pinned
-    /// crop, which are floating over exactly the area being photographed.
-    func hideOwnWindows() -> [NSWindow] {
-        // Counted, because a capture can start while a pinned crop is still
-        // choosing its region: the first one to finish must not put the
-        // overlays back into the other one's photograph.
-        captureDepth += 1
-        mouse.suspend()
-        let hidden = (presenter.visible + crops.visibleWindows + [guidePanel].compactMap { $0 })
-            .filter { $0.isVisible }
-        hidden.forEach { $0.orderOut(nil) }
-        return hidden
-    }
-
-    func showOwnWindows(_ hidden: [NSWindow]) {
-        captureDepth = max(0, captureDepth - 1)
-        hidden.forEach { $0.orderFront(nil) }
-        guard captureDepth == 0 else { return }
-        mouse.resume()
-    }
-
-    private func finishShot(copied: Bool, path: String?) {
-        if let path, store.config.shotCopyToClipboard, !copied,
-           let saved = NSImage(contentsOfFile: path) {
-            ScreenshotManager.copyToClipboard(saved)
-        }
-        model.shotStatus = String(localized: path.map { .hudSavedTo($0) } ?? .hudCopiedToClipboard)
-    }
-
-    func openPage(_ id: String) {
-        model.selectedPage = id
-        openMainWindow()
-    }
-}
-
-// MARK: - AppActions
-
-extension AppDelegate: AppActions {
-
-    /// The one path a setting takes from the UI to disk. `ConfigStore` clamps
-    /// and saves it, `applyConfig` hands it to whoever needs it. A new setting
-    /// therefore adds nothing here, and cannot be saved but left unapplied,
-    /// which is the mistake a method per setting made easy to make.
-    func update<Value>(_ path: WritableKeyPath<Config, Value>, to value: Value) {
-        store.update { $0[keyPath: path] = value }
-    }
-
-    /// Keep-awake is held by its manager, not by a stored flag: a hold made by
-    /// hand outlives the schedule that would otherwise decide.
-    func setAwake(_ on: Bool) { awake.set(on) }
-
-    func setHotkey(_ action: HotkeyAction, to hotkey: Hotkey) {
-        // Taking a combination frees it wherever it was, possibly on a page
-        // the user is not looking at. Config.bind is where that rule lives.
-        var stolen: [HotkeyAction] = []
-        store.update { stolen = $0.bind(action, to: hotkey) }
-        if let taken = stolen.first {
-            HUD.shared.show(.hudHotkeyTaken(hotkey.display, String(localized: taken.title)))
-        }
-    }
-
-    func clearHotkey(_ action: HotkeyAction) {
-        update(\.hotkeys[action.rawValue], to: "")
-    }
-
-    func resetHotkeys() {
-        update(\.hotkeys, to: [:])
-    }
-
-    func setLaunchAtLogin(_ on: Bool) {
-        applyLaunchAtLogin(on)
-        update(\.launchAtLogin, to: on)
-        // macOS can refuse, so the toggle shows what it settled on rather than
-        // what it was asked for.
-        model.launchAtLogin = isLaunchAtLoginEnabled
-    }
-
-    func capture(_ mode: CaptureMode) {
-        runCapture(mode, openEditor: true)
-    }
-
-    func flashZones() {
-        dragSnap.previewZones()
-    }
-
-    func reportBug() {
-        guard let url = URL(string: Self.issuesURL) else { return }
-        NSWorkspace.shared.open(url)
-    }
-
-    func launchWorkspace(named name: String, onScreen screen: Int?) {
-        guard let workspace = store.config.workspaces[name] else { return }
-        launcher.launch(workspace, named: name, onScreen: screen)
-    }
-
-    func deleteWorkspace(named name: String) {
-        store.update { $0.workspaces.removeValue(forKey: name) }
-        refreshWorkspaceModel()
-    }
-
-    func renameWorkspace(_ old: String, to new: String) -> Bool {
-        guard old != new else { return true }
-        guard store.config.workspaces[old] != nil, store.config.workspaces[new] == nil else { return false }
-        store.update {
-            guard let workspace = $0.workspaces.removeValue(forKey: old) else { return }
-            $0.workspaces[new] = workspace
-        }
-        refreshWorkspaceModel()
-        return true
-    }
-
-    func saveCurrentWorkspace(named name: String) {
-        let items = router.snapshotWorkspace()
-        guard !items.isEmpty else { return }
-        let moveExisting = store.config.workspaces[name]?.moveExisting ?? true
-        store.update { $0.workspaces[name] = Workspace(items: items, moveExisting: moveExisting) }
-        refreshWorkspaceModel()
-    }
-
-    func setWorkspaceMoveExisting(_ on: Bool, for name: String) {
-        store.update { $0.workspaces[name]?.moveExisting = on }
-        refreshWorkspaceModel()
-    }
-
-    func updateWorkspaceItem(_ index: Int, in name: String, urls: [String], args: [String]) {
-        store.update {
-            guard $0.workspaces[name]?.items.indices.contains(index) == true else { return }
-            $0.workspaces[name]?.items[index].urls = urls.isEmpty ? nil : urls
-            $0.workspaces[name]?.items[index].args = args.isEmpty ? nil : args
-        }
-        refreshWorkspaceModel()
-    }
-
-    func removeWorkspaceItem(_ index: Int, from name: String) {
-        store.update {
-            guard $0.workspaces[name]?.items.indices.contains(index) == true else { return }
-            $0.workspaces[name]?.items.remove(at: index)
-        }
-        refreshWorkspaceModel()
-    }
-
-    func cancelWorkspaceLaunch() {
-        launcher.cancel()
-    }
-
-    func checkForUpdates() {
-        updates.check()
-    }
-
-    func installUpdate() {
-        updates.install()
-    }
-
-    func openReleasePage() {
-        NSWorkspace.shared.open(updates.available?.pageURL ?? Release.releasesPageURL)
     }
 }
