+++ scripts/line-limit-baseline
@@ -8,7 +8,6 @@
 # delete the line once it is under the limit.
 #
 # lines  path
-867   App/Sources/plonk/AppDelegate.swift
 489   App/Tests/plonkTests/ConfigTests.swift
 444   App/Tests/plonkTests/RouterTests.swift
 428   App/Sources/plonk/WindowManager.swift
